gen-ai/Assistants/law_firm_assistant/requirements.txt
openai==1.30.1
python-dotenv==0.21.0
